# swagger/openapi.yaml
openapi: 3.0.3
info:
  title: Example API
  description: API documentation for the backend service
  version: 1.0.0
servers:
  - url: http://backend.vm1.test/api
    description: Development server
paths:
...
  /users/{id}:
    get:
      summary: Get a single user
      operationId: getUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User details
    put:
      summary: Update a user
      operationId: updateUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: User updated
    delete:
      summary: Delete a user
      operationId: deleteUser
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: User deleted